Introduced Error Handling, Metaprogramming of find_by methods

+++ lib/udacidata.rb
@@ -65,41 +65,49 @@ class Udacidata
 	def self.find(n)
 		products = CSV.read(@@data_path)
 		product_n = products[n]
-		return @found_product = Product.new(id: product_n[0].to_i, brand: product_n[1], name: product_n[2], price:product_n[3].to_f)
+		if n <= products.length - 1 
+			return @found_product = Product.new(id: product_n[0].to_i, brand: product_n[1], name: product_n[2], price:product_n[3].to_f)
+		else 
+			raise Errors::ProductNotFoundError, "This item couldn't be found in the database"
+		end
 	end
 
 	def self.destroy(n)
 		products = CSV.read(@@data_path)
 		deleted_product = products[n]
 		products.delete_at(n)
-		products.shift
-		File.delete(@@data_path)
-		CSV.open(@@data_path, "ab") do |csv|
-			csv << ["id", "brand", "product", "price"]
-			products.each do |product|
-			csv << [product[0], product[1], product[2], product[3].to_f]
-		 	end
-		end
-		return @destroyed_product = Product.new(id: deleted_product[0].to_i, brand: deleted_product[1], name: deleted_product[2], price:deleted_product[3].to_f)
-	end
-
-	def self.find_by_brand(brand)
-		products = CSV.read(@@data_path)
-		products.each do |product|
-			if product[1] == brand
-				return @found_product = Product.new(id: product[0], brand: product[1], name: product[2], price: product[3].to_f)
+		if n <= products.length - 1 
+			products.shift
+			File.delete(@@data_path)
+			CSV.open(@@data_path, "ab") do |csv|
+				csv << ["id", "brand", "product", "price"]
+				products.each do |product|
+				csv << [product[0], product[1], product[2], product[3].to_f]
+			 	end
 			end
+			return @destroyed_product = Product.new(id: deleted_product[0].to_i, brand: deleted_product[1], name: deleted_product[2], price:deleted_product[3].to_f)
+		else
+			raise Errors::ProductNotFoundError, "This item couldn't be found in the database"
 		end
 	end
 
-	def self.find_by_name(name)
-		products = CSV.read(@@data_path)
-		products.each do |product|
-			if product[2] == name
-				return @found_product = Product.new(id: product[0], brand: product[1], name: product[2], price: product[3].to_f)
-			end
-		end
-	end
+	#def self.find_by_brand(brand)
+	#	products = CSV.read(@@data_path)
+	#	products.each do |product|
+	#		if product[1] == brand
+	#			return @found_product = Product.new(id: product[0], brand: product[1], name: product[2], price: product[3].to_f)
+	#		end
+	#	end
+	#end
+
+	#def self.find_by_name(name)
+	#	products = CSV.read(@@data_path)
+	#	products.each do |product|
+	#		if product[2] == name
+	#			return @found_product = Product.new(id: product[0], brand: product[1], name: product[2], price: product[3].to_f)
+	#		end
+	#	end
+	#end
 
 	def self.where(options={})
 		brand = options[:brand]
@@ -124,7 +132,6 @@ class Udacidata
 		updated_product = Product.new(id: self.id, brand: new_brand == nil ? self.brand : new_brand, name: new_name == nil ? self.name : new_name, price: new_price == nil ? self.price : new_price)
 		updated_product_arr = [updated_product.id, updated_product.brand, updated_product.name, updated_product.price]
 		products[self.id - 1] = updated_product_arr
-		p products
 
 		File.delete(@@data_path)
 		CSV.open(@@data_path, "ab") do |csv|
